Transpose outputs of gbm_to_params
All of the arrays provided by gbm_to_params had trees as the first
axis. However, braintree models turned out to be much easier to
implement with terminal nodes as the first axis. Swapped the axes of
the outputs of gbm_to_params so they can be directly copied into the
variables of a braintree without tranposition.