Initialize braintree with xgboost parameters
Refactored xgboost fitting and parameter translation to be part of an
initialize() method belonging to braintree. This makes sense for
several reasons, not the least of which is the fact that the xgboost
fit can share the max_depth, num_trees, and num_features parameters
with the braintree model.

In the process of figuring out why the xgboost initialization didn't
immediately help the braintree model scores, I found and fixed a
couple of bugs in the original braintree scoring implementation. Some
terminal nodes were never getting scored.