Paramterize noise levels of weight matrices
Added parameters to the braintree constructor to control the amount of
Gaussian noise present in the split and terminal weight matrices when
they are initialized. There is reason to think that these values will
affect model performance, so it's worth exposing them as parameters.

The values the users supplies are actually the standard deviation of
the sum of the columns. This parameterization allows for the model to
automatically reduce the standard deviation when the number of features
increases, in line with deep learning best practices.